Fix some compiler warnings

+++ src/util/summarize.cpp
@@ -32,7 +32,7 @@ int summarize(int argc, const char** argv) {
     Debug(Debug::INFO) << "Start writing to file " << par.db3 << "\n";
 
     #pragma omp for schedule(dynamic, 100)
-    for (int i = 0; i < reader.getSize(); ++i) {
+    for (size_t i = 0; i < reader.getSize(); ++i) {
         int thread_idx = 0;
 #ifdef OPENMP
         thread_idx = omp_get_thread_num();
